src/requirements.txt
autopep8==2.0.2
black==23.1.0
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
platformdirs==3.1.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
